requirements.txt
certifi==2019.11.28
click==7.1.1
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
py7zr==0.6
pycryptodome==3.9.7
PyPDF2==1.26.0
python-docx==0.8.10
texttable==1.6.2
Werkzeug==1.0.0
wincertstore==0.2
transliterate==1.10.2